ASAA-10 Moved data classes that were states to individual files in the same package
ASAA-10 refactored out UI models from other compose components so that they are decoupled
