+++ compose/src/main/java/com/bottlerocketstudios/compose/appbar/ArchAppBar.kt
@@ -16,8 +16,6 @@ import androidx.navigation.NavHostController
 import com.bottlerocketstudios.compose.widgets.AppBar
 import kotlinx.coroutines.launch
 
-
-
 @Composable
 fun ArchAppBar(
     state: ArchAppBarState,
@@ -58,9 +56,3 @@ fun ArchAppBar(
         )
     }
 }
-
-data class ArchAppBarState(
-    val showToolbar: State<Boolean>,
-    val topLevel: State<Boolean>,
-    val title: State<String>,
-)
