+++ compose/src/main/java/com/bottlerocketstudios/compose/appbar/ArchAppBarState.kt
@@ -1,0 +1,9 @@
+package com.bottlerocketstudios.compose.appbar
+
+import androidx.compose.runtime.State
+
+data class ArchAppBarState(
+    val showToolbar: State<Boolean>,
+    val topLevel: State<Boolean>,
+    val title: State<String>,
+)
