+++ compose/src/main/java/com/bottlerocketstudios/compose/auth/AuthCodeScreen.kt
@@ -152,13 +152,5 @@ private fun AuthCodeContent(state: AuthCodeState) {
     }
 }
 
-data class AuthCodeState(
-    val requestUrl: State<String>,
-    val devOptionsEnabled: Boolean,
-    val onAuthCode: (String) -> Unit,
-    val onLoginClicked: () -> Unit,
-    val onSignupClicked: () -> Unit,
-    val onDevOptionsClicked: () -> Unit,
-    val showToolbar: (show: Boolean) -> Unit,
-)
+
