+++ compose/src/main/java/com/bottlerocketstudios/compose/auth/LoginScreen.kt
@@ -301,16 +301,3 @@ private fun PreviewOuterScreenContent() {
         )
     }
 }
-
-data class LoginScreenState(
-    val email: State<String>,
-    val password: State<String>,
-    val loginEnabled: State<Boolean>,
-    val devOptionsEnabled: Boolean,
-    val onLoginClicked: () -> Unit,
-    val onSignupClicked: () -> Unit,
-    val onForgotClicked: () -> Unit,
-    val onDevOptionsClicked: () -> Unit,
-    val onEmailChanged: (String) -> Unit,
-    val onPasswordChanged: (String) -> Unit
-)
