+++ compose/src/main/java/com/bottlerocketstudios/compose/home/HomeScreen.kt
@@ -83,8 +83,3 @@ fun HomeScreenPreview() {
         )
     }
 }
-
-data class HomeScreenState(
-    val repositories: State<List<UserRepositoryUiModel>>,
-    val itemSelected: (userRepositoryUiModel: UserRepositoryUiModel) -> Unit,
-)
