+++ compose/src/main/java/com/bottlerocketstudios/compose/home/HomeScreenState.kt
@@ -1,0 +1,8 @@
+package com.bottlerocketstudios.compose.home
+
+import androidx.compose.runtime.State
+
+data class HomeScreenState(
+    val repositories: State<List<UserRepositoryUiModel>>,
+    val itemSelected: (userRepositoryUiModel: UserRepositoryUiModel) -> Unit,
+)
