+++ compose/src/main/java/com/bottlerocketstudios/compose/navdrawer/NavDrawerItem.kt
@@ -92,9 +92,4 @@ fun NavItemSelectedPreview() {
     }
 }
 
-data class NavItemState(
-    @DrawableRes val icon: Int,
-    @StringRes val itemText: Int,
-    val selected: Boolean,
-    val onClick: suspend () -> Unit
-)
+
