+++ compose/src/main/java/com/bottlerocketstudios/compose/navdrawer/NavDrawerState.kt
@@ -1,0 +1,11 @@
+package com.bottlerocketstudios.compose.navdrawer
+
+import androidx.compose.runtime.State
+
+data class NavDrawerState(
+    val avatarUrl: State<String>,
+    val displayName: State<String>,
+    val username: State<String>,
+    val items: State<List<NavItemState>>,
+    val devOptionsListener: () -> Unit,
+)
