+++ compose/src/main/java/com/bottlerocketstudios/compose/repository/FileBrowserScreenState.kt
@@ -1,0 +1,8 @@
+package com.bottlerocketstudios.compose.repository
+
+import androidx.compose.runtime.State
+
+data class FileBrowserScreenState(
+    val path: State<String>,
+    val content: State<ByteArray?>,
+)
