+++ compose/src/main/java/com/bottlerocketstudios/compose/repository/RepositoryBrowserScreenState.kt
@@ -1,0 +1,10 @@
+package com.bottlerocketstudios.compose.repository
+
+import androidx.compose.runtime.State
+
+data class RepositoryBrowserScreenState(
+    val path: State<String>,
+    val itemCount: State<Int>,
+    val repositoryItems: State<List<RepositoryItemUiModel>>,
+    val onRepositoryItemClicked: (RepositoryItemUiModel) -> Unit
+)
