+++ compose/src/main/java/com/bottlerocketstudios/compose/util/StringIdHelper.kt
@@ -40,7 +40,7 @@ sealed class StringIdHelper : DomainModel, Serializable, KoinComponent {
                     // Allow for the use of string helpers within format args by unwrapping them here
                     if (it is StringIdHelper) it.getString() else it
                 }
-                context.resources.getQuantityString(idRes, quantity, mappedArgs.toTypedArray())
+                context.resources.getQuantityString(idRes, quantity, *mappedArgs.toTypedArray())
             }
         }
     }
