+++ domain/src/main/java/com/bottlerocketstudios/brarchitecture/domain/models/Status.kt
@@ -81,7 +81,7 @@ fun Exception.asAppropriateFailure(): Status.Failure = when (this) {
  */
 // FIXME- make logger interface for domain
 suspend fun <T : Any> logWrappedExceptions(className: String, methodName: String, block: suspend () -> Status<T>): Status<T> {
-    @Suppress("TooGenericExceptionCaught") // this might be overly broad (catching NPEs isn't ideal) but not worth it yet to check for all other possible types (IO..., Json..., others)
+    @Suppress("TooGenericExceptionCaught") // this might be overly broad (catching general Exceptions isn't ideal) but not worth it yet to check for all other possible types (IO..., Json..., others)
     return try {
         // Timber.tag(className).v("[$methodName]")
         block().also {
@@ -96,5 +96,5 @@ suspend fun <T : Any> logWrappedExceptions(className: String, methodName: String
 
 // Simple version of wrap exceptions; with no logging.
 suspend fun <T : Any> wrapExceptions(block: suspend () -> Status<T>): Status<T> =
-    @Suppress("TooGenericExceptionCaught") // this might be overly broad (catching NPEs isn't ideal) but not worth it yet to check for all other possible types (IO..., Json..., others)
+    @Suppress("TooGenericExceptionCaught") // this might be overly broad (catching general Exceptions isn't ideal) but not worth it yet to check for all other possible types (IO..., Json..., others)
     try { block() } catch (e: Exception) { e.asAppropriateFailure() }
